# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  models.DeveloperReq:
    properties:
      developer_role:
        type: string
      employee_id:
        type: string
    type: object
  models.DeveloperRes:
    properties:
      developer_role:
        type: string
      employee_id:
        type: string
      id:
        type: string
    type: object
  models.EmployeeList:
    properties:
      employees:
        items:
          $ref: '#/definitions/models.EmployeeRes'
        type: array
    type: object
  models.EmployeeReq:
    properties:
      birth_date:
        example: 2006/01/02
        type: string
      full_name:
        type: string
      phone:
        type: string
      position:
        type: string
      profile_photo:
        type: string
      role:
        type: string
    type: object
  models.EmployeeRes:
    properties:
      birth_date:
        example: 2006/01/02
        type: string
      full_name:
        type: string
      id:
        type: string
      phone:
        type: string
      position:
        type: string
      profile_photo:
        type: string
      role:
        type: string
    type: object
  models.FailureInfo:
    properties:
      description:
        type: string
      error: {}
      status_code:
        type: integer
    type: object
...
  models.GetEmployee:
    properties:
      access_token:
        type: string
      birth_date:
        type: string
      full_name:
        type: string
      id:
        type: string
      phone:
        type: string
      position:
        type: string
      profile_photo:
        type: string
      role:
        type: string
    type: object
...
  models.Success:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
...
host: localhost:7777
info:
  contact:
    email: gofurovmurtazoxon@gmail.com
    name: Murtazoxon
    url: https://t.me/murtazokhon_gofurov
  description: This is Task api server
  termsOfService: 2 term OSG
  title: Online Service Group Task api
  version: "1.0"
paths:
  /developer:
    post:
      consumes:
      - application/json
      description: Through this api, can create a developer
      parameters:
      - description: CreateRole
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.DeveloperReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DeveloperRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Create an employee
      tags:
      - Developer
  /employee:
    post:
      consumes:
      - application/json
      description: Through this api, can create an employee
      parameters:
      - description: CreateEmployee
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.EmployeeReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Create new employee
      tags:
      - Employee
    put:
      consumes:
      - application/json
      description: Through this api, can UPDATE an employee
      parameters:
      - description: UpdateEmployee
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.EmployeeRes'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmployeeRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: UPDATE an employee
      tags:
      - Employee
  /employee/{id}:
    delete:
      consumes:
      - application/json
      description: Through this api, can Delete an employee
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: DELETE an employee
      tags:
      - Employee
    get:
      consumes:
      - application/json
      description: Through this api, can GET an employee
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: GET an employee
      tags:
      - Employee
  /employees:
    get:
      consumes:
      - application/json
      description: Through this api, can GET employees
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmployeeList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: GET employees
      tags:
      - Employee
...
  /login/{phone}:
    get:
      consumes:
      - application/json
      description: Through this api, can login an employee
      parameters:
      - description: phone_number
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetEmployee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: Login an employee
      tags:
      - Login
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"